+++ blog-after/src/main/java/com/example/demo/service/impl/UserServiceImpl.java
@@ -18,7 +18,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findUserById(Integer id) {
+    public User findUserById(Long id) {
         return userMapper.findUserById(id);
     }
 }
